+++ internal/pkg/bifs/relative_time.go
@@ -12,7 +12,7 @@ func BIF_dhms2sec(input1 *mlrval.Mlrval) *mlrval.Mlrval {
 	if !input1.IsString() {
 		return mlrval.ERROR
 	}
-	var d, h, m, s int
+	var d, h, m, s int64
 
 	if strings.HasPrefix(input1.AcquireStringValue(), "-") {
 
@@ -113,7 +113,7 @@ func BIF_hms2sec(input1 *mlrval.Mlrval) *mlrval.Mlrval {
 	if input1.AcquireStringValue() == "" {
 		return mlrval.ERROR
 	}
-	var h, m, s int
+	var h, m, s int64
 
 	if strings.HasPrefix(input1.AcquireStringValue(), "-") {
 		n, err := fmt.Sscanf(input1.AcquireStringValue(), "-%d:%d:%d", &h, &m, &s)
@@ -159,7 +159,7 @@ func BIF_sec2dhms(input1 *mlrval.Mlrval) *mlrval.Mlrval {
 		return mlrval.ERROR
 	}
 
-	var d, h, m, s int
+	var d, h, m, s int64
 
 	splitIntToDHMS(isec, &d, &h, &m, &s)
 	if d != 0 {
@@ -194,7 +194,7 @@ func BIF_sec2hms(input1 *mlrval.Mlrval) *mlrval.Mlrval {
 		isec = -isec
 	}
 
-	var d, h, m, s int
+	var d, h, m, s int64
 
 	splitIntToDHMS(isec, &d, &h, &m, &s)
 	h += d * 24
@@ -212,15 +212,15 @@ func BIF_fsec2dhms(input1 *mlrval.Mlrval) *mlrval.Mlrval {
 		return mlrval.ERROR
 	}
 
-	sign := 1
+	sign := int64(1)
 	if fsec < 0 {
 		sign = -1
 		fsec = -fsec
 	}
-	isec := int(math.Trunc(fsec))
+	isec := int64(math.Trunc(fsec))
 	fractional := fsec - float64(isec)
 
-	var d, h, m, s int
+	var d, h, m, s int64
 
 	splitIntToDHMS(isec, &d, &h, &m, &s)
 
@@ -267,10 +267,10 @@ func BIF_fsec2hms(input1 *mlrval.Mlrval) *mlrval.Mlrval {
 		sign = "-"
 		fsec = -fsec
 	}
-	isec := int(math.Trunc(fsec))
+	isec := int64(math.Trunc(fsec))
 	fractional := fsec - float64(isec)
 
-	var d, h, m, s int
+	var d, h, m, s int64
 
 	splitIntToDHMS(isec, &d, &h, &m, &s)
 	h += d * 24
@@ -290,12 +290,12 @@ func BIF_fsec2hms(input1 *mlrval.Mlrval) *mlrval.Mlrval {
 }
 
 // Helper function
-func splitIntToDHMS(u int, pd, ph, pm, ps *int) {
-	d := 0
-	h := 0
-	m := 0
-	s := 0
-	sign := 1
+func splitIntToDHMS(u int64, pd, ph, pm, ps *int64) {
+	d := int64(0)
+	h := int64(0)
+	m := int64(0)
+	s := int64(0)
+	sign := int64(1)
 	if u < 0 {
 		u = -u
 		sign = -1
